feat: Add a configuration manager
