+++ merino/main.py
@@ -5,7 +5,7 @@ from fastapi.exceptions import RequestValidationError
 from fastapi.responses import JSONResponse
 
 from merino import providers
-from merino.config import configure_logging
+from merino.config_logging import configure_logging
 from merino.middleware import logging
 from merino.web import api_v1, dockerflow
