+++ pyproject.toml
@@ -42,6 +42,7 @@ kinto-http = "^10.10.1"
 requests = "^2.28.1"
 dockerflow = "^2022.7.0"
 asgi-correlation-id = "^3.0.1"
+dynaconf = "^3.1.9"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.1.2"
